+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added `internal/credentials.IsAccessError(err)` helper for check access errors
 * Changed period for re-fresh static credentials token from `1/2` to `1/10` to expiration time
 
 ## v3.53.4